add  modal react library refactor theme select config file
