+++ package.json
@@ -26,10 +26,11 @@
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.27.0",
     "react-i18next": "^11.15.5",
+    "react-modal": "^3.14.4",
     "react-redux": "^7.2.6",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.2.2",
     "redux-logger": "^3.0.6",
     "typescript": "^4.5.5",
