+++ src/config/modal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './modal-config';
