+++ src/config/modal/modal-config.ts
@@ -1,0 +1,12 @@
+const modalConfig = {
+  content: {
+    top: '50%',
+    left: '50%',
+    right: 'auto',
+    bottom: 'auto',
+    marginRight: '-50%',
+    transform: 'translate(-50%, -50%)',
+  },
+};
+
+export default modalConfig;
