+++ src/config/react-select/index.ts
@@ -1,1 +1,1 @@
-export { default } from './theme';
+export { default } from './theme-config';
